muka-api/requirements.txt
Flask==2.2.2
tensorflow==2.11.0
google-cloud-storage==2.9.0
Pillow==9.3.0
